nova_nexus/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
google-generativeai==0.7.2
pydantic==2.11.10
python-dotenv==1.0.1
